+++ matplotlive/sketch.py
@@ -54,9 +54,16 @@ class Sketch:
         self.lines = {}
         self.right_axis = right_axis
 
-    def redraw(self):
+    def redraw(self) -> None:
         """Redraw the entire plot (e.g. after updating axis labels)."""
         plt.show(block=False)
+
+    def reset(self) -> None:
+        """Reset the sketch."""
+        for line in self.lines.values():
+            line.remove()
+        self.lines = {}
+        self.update()
 
     def add_line(self, name: str, side: str, *args, **kwargs) -> None:
         """Add a line-plot to the left axis.
@@ -92,11 +99,11 @@ class Sketch:
         """
         self.lines[name].set_data(xdata, ydata)
 
-    def __draw_lines(self):
+    def __draw_lines(self) -> None:
         for line in self.lines.values():
             self.figure.draw_artist(line)
 
-    def __on_draw(self, event):
+    def __on_draw(self, event) -> None:
         if event is not None:
             if event.canvas != self.canvas:
                 raise RuntimeError