+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- LivePlot: Add a `reset` function
+- Sketch: Add a `reset` function
+
 ## [1.0.0] - 2024-08-06
 
 ### Added